More work. Extracted WebSocketApplication.kt from Application.kt (#80)
